[refactor] create_communicatorとcreate_jwt_for_userをconftest.pyに集約
